# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Admin API for managing portfolio content",
        "title": "Portfolio Admin API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8083",
    "basePath": "/api/v1",
    "paths": {
...
        "/portfolio/projects": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all portfolio projects",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portfolio - Projects"
                ],
                "summary": "Get all portfolio projects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_GunarsK-portfolio_admin-api_internal_models.PortfolioProject"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new portfolio project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portfolio - Projects"
                ],
                "summary": "Create portfolio project",
                "parameters": [
                    {
                        "description": "Portfolio project data",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_GunarsK-portfolio_admin-api_internal_models.PortfolioProject"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_GunarsK-portfolio_admin-api_internal_models.PortfolioProject"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/portfolio/projects/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a single portfolio project by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portfolio - Projects"
                ],
                "summary": "Get portfolio project by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Portfolio Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_GunarsK-portfolio_admin-api_internal_models.PortfolioProject"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing portfolio project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portfolio - Projects"
                ],
                "summary": "Update portfolio project",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Portfolio Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Portfolio project data",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_GunarsK-portfolio_admin-api_internal_models.PortfolioProject"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_GunarsK-portfolio_admin-api_internal_models.PortfolioProject"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a portfolio project and all associated technology links\nNote: This is a cascade delete - one API call deletes the project and all technology associations\nThe project image file in S3 is preserved and cleaned up by background job",
                "tags": [
                    "Portfolio - Projects"
                ],
                "summary": "Delete portfolio project",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Portfolio Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_GunarsK-portfolio_admin-api_internal_models.PortfolioProject": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "challenges": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "displayOrder": {
                    "type": "integer"
                },
                "endDate": {
                    "type": "string"
                },
                "featured": {
                    "type": "boolean"
                },
                "features": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "githubUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imageFile": {
                    "description": "Associations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_GunarsK-portfolio_portfolio-common_models.StorageFile"
                        }
                    ]
                },
                "imageFileId": {
                    "type": "integer"
                },
                "isOngoing": {
                    "type": "boolean"
                },
                "learnings": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "liveUrl": {
                    "type": "string"
                },
                "longDescription": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "teamSize": {
                    "type": "integer"
                },
                "technologies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_GunarsK-portfolio_portfolio-common_models.Skill"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "github_com_GunarsK-portfolio_portfolio-common_models.Skill": {
            "type": "object",
            "required": [
                "skill",
                "skillTypeId"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "displayOrder": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isVisible": {
                    "type": "boolean"
                },
                "skill": {
                    "type": "string"
                },
                "skillType": {
                    "$ref": "#/definitions/github_com_GunarsK-portfolio_portfolio-common_models.SkillType"
                },
                "skillTypeId": {
                    "type": "integer"
                },
                "type": {
                    "description": "Computed field",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_GunarsK-portfolio_portfolio-common_models.SkillType": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "displayOrder": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_GunarsK-portfolio_portfolio-common_models.StorageFile": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "fileSize": {
                    "type": "integer"
                },
                "fileType": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mimeType": {
                    "type": "string"
                },
                "url": {
                    "description": "Computed field",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}